Moved MIG interface to separate file
